Início da adição dos DTOs

+++ src/main/java/com/igor_dev/GreenTrace/service/VendaService.java
@@ -1,14 +1,17 @@
 package com.igor_dev.GreenTrace.service;
 
+import java.time.OffsetDateTime;
 import java.util.List;
-import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import com.igor_dev.GreenTrace.model.Funcionarios;
+import com.igor_dev.GreenTrace.dto.VendaRequestDTO;
+import com.igor_dev.GreenTrace.dto.VendaResponseDTO;
 import com.igor_dev.GreenTrace.model.Vendas;
 import com.igor_dev.GreenTrace.repository.VendaRepository;
+
+import jakarta.persistence.EntityNotFoundException;
 
 @Service
 public class VendaService {
@@ -16,34 +19,62 @@ public class VendaService {
     @Autowired
     private VendaRepository vendaRepository;
     
-        public Vendas createVenda(Funcionarios funcionario, Vendas venda){
-        return vendaRepository.save(venda);
+    public VendaResponseDTO toResponse(Vendas venda){
+        return new VendaResponseDTO(
+            venda.getId_venda(),
+            venda.getCliente(),
+            venda.getProduto(),
+            venda.getFuncionario(),
+            venda.getCriadoEm()
+        );
     }
 
-    public List<Vendas> findAllVenda(){
-        return vendaRepository.findAll();
-    }
-
-    public Optional<Vendas> findVendaById(Long id){
-        return vendaRepository.findById(id);
-    }
-
-
-    public Vendas updateVenda (Long id, Vendas venda){
-        Vendas vendaEntity = findVendaById(id)
-        .orElseThrow(() -> new RuntimeException("Produto não encontrado"));
-        
-        Vendas vendaUpdated = Vendas.builder()
-        .id_venda(id)
-        .cliente(venda.getCliente() != null ? venda.getCliente() : vendaEntity.getCliente())
-        .produto(venda.getProduto() != null ? venda.getProduto() : vendaEntity.getProduto())
-        .funcionario(venda.getFuncionario() != null ? venda.getFuncionario() : vendaEntity.getFuncionario())
+    public VendaResponseDTO criarVenda(VendaRequestDTO vendaRequestDTO) {
+        Vendas venda = Vendas.builder()
+        .cliente(vendaRequestDTO.cliente())
+        .produto(vendaRequestDTO.produto())
+        .funcionario(vendaRequestDTO.funcionario())
+        .criadoEm(OffsetDateTime.now())
         .build();
 
-        return vendaRepository.save(vendaUpdated);
+        return toResponse(vendaRepository.save(venda));
+    }
+
+
+
+    public VendaResponseDTO findVendaById(Long id){
+        Vendas venda = vendaRepository.findById(id).orElseThrow(() -> new EntityNotFoundException("Não foi possível encontrar a venda com o id: " + id));
+        return toResponse(venda);
+    }
+
+    
+    public VendaResponseDTO findAllVenda(){
+        List<Vendas> vendas = vendaRepository.findAll();
+        return toResponse((Vendas) vendas);
     }
 
     public void deletarVenda(Long id){
+        if(!vendaRepository.existsById(id)){
+            throw new EntityNotFoundException("Venda não encontrada");
+        }
+        
         vendaRepository.deleteById(id);
     }
+
+
+    // TODO
+    // public Vendas updateVenda (Long id, Vendas venda){
+    //     Vendas vendaEntity = findVendaById(id);
+        
+    //     Vendas vendaUpdated = Vendas.builder()
+    //     .id_venda(id)
+    //     .cliente(venda.getCliente() != null ? venda.getCliente() : vendaEntity.getCliente())
+    //     .produto(venda.getProduto() != null ? venda.getProduto() : vendaEntity.getProduto())
+    //     .funcionario(venda.getFuncionario() != null ? venda.getFuncionario() : vendaEntity.getFuncionario())
+    //     .build();
+
+    //     return toResponse(vendaRepository.save(vendaUpdated));
+    // }
+
+
 }
